metrics: Ignore old clients
The metrics protocol went through multiple versions.

The current one is version 2, which is in fact what already was on the
first public releases of Endless OS.

As a result, we aren't interested at all in metrics requests version 0
and 1, they just won't provide interesting data.

Since we're only handling version 2, we can simplify the code a bit as
well.