redis: Support SSL connections
If you connect to an SSL enabled Redis server, `create_redis_pool` will
hang forever giving the impression that the proxy is running. While it's
possible to run a TLS proxy for this configuration, aioredis supports
SSL wrapping the TCP connection out of the box[1]. Expose this as an
`ssl` boolean in the configuration. While here, add a log message after
the connection is made to help give some indication that the proxy isn't
hung there.

1. https://aioredis.readthedocs.io/en/v1.3.1/api_reference.html#aioredis.create_redis_pool