docker: Make commands work correctly out of the box
If the `-c /tmp/config.toml` is in the `CMD`, then the documented
suggestion of using `proxy <subcommand>` as the container command will
not reflect the generated configuration. Move that into the `ENTRYPOINT`
so that the container command is a server subcommand.

This is how azafea is setup, and I think it's the most natural way. It
does mean that if you want to run a separate pipenv script or command
that you have to override the entrypoint. That's unfortunate but I think
that's fairly niche as seen in the GitHub workflow.