Require python 3.8
This matches what's currently used in Azafea. It would be better to 3.11
matching stable Endless OS and Debian, but that brought in too much
breakage and will need to wait for another day. It also would have been
possible to stick with 3.7, but many of the dependencies have dropped
support for it and trying to operate pipenv using the 3.7 from pyenv
fails for me. Just use 3.8 for now.